reworked routing; changed nav names; reinstated shop as a page; renamed shopItems to merch
