+++ client/src/App.js
@@ -7,7 +7,11 @@ import Home from './pages/Home.js';
 import Signup from './pages/Signup.js';
 import Login from './pages/Login.js';
 import Dashboard from './pages/Dashboard.js';
-import Shop from './components/Shop';
+import Shop from './pages/Shop.js';
+
+import Artists from './components/Artists';
+import ArtistPortfolio from './components/ArtistPortfolio';
+import ArtistShop from './components/ArtistShop';
 
 // import PostDetails from './components/PostDetails';
 
@@ -23,24 +27,20 @@ function App() {
     <ApolloProvider client={client}>
       <BrowserRouter>
         <GlobalHeader />
-        {/* prettier-ignore */}
         <Routes>
           <Route path="/" element={<Home />} />
-            <Route path="signup" element={<Signup />} />
-            <Route path="login" element={<Login />} />
+          <Route path="dashboard" element={<Dashboard />} />
+          <Route path="dashboard/my/table" element={<Shop />} />
 
-            <Route path="dashboard" element={<Dashboard />} />
-              <Route path="dashboard/my/table" element={<Shop />} />
+          {/* <Route path="posts/:postId" element={<PostDetails/>}/> */}
 
-            {/* <Route path="posts/:postId" element={<PostDetails/>}/> */}
+          {/* need an index element: artists w/ outlet to display artist profile page & artist store page */}
+          <Route path="artists" element={<Artists />} />
+          <Route path="artists/someArtist" element={<ArtistPortfolio />} />
+          <Route path="artists/someArtist/table" element={<ArtistShop />} />
 
-{/* need an index element: artists w/ outlet to display artist profile page & artist store page */}
-            <Route path="artists"/>
-              {/* 
-              <Route path=":artistId" element={<ArtistProfile />} />
-              <Route path=":artistId/table" element={<ArtistShop} /> /> 
-              */}
-              {/* <Route index element={<Artists/>} /> */}
+          <Route path="/signup" element={<Signup />} />
+          <Route path="/login" element={<Login />} />
         </Routes>
       </BrowserRouter>
     </ApolloProvider>
