+++ client/src/components/ArtistPortfolio/index.js
@@ -2,10 +2,10 @@ import './index.css';
 
 const ArtistPortfolio = () => {
   return (
-    <>
+    <div>
       <h2>this is the artist's portfolio page</h2>
       YOMMY
-    </>
+    </div>
   );
 };
