+++ client/src/components/ArtistShop/index.js
@@ -1,0 +1,12 @@
+import './index.css';
+
+const ArtistShop = () => {
+  return (
+    <>
+      <h2>this is the artist's shop page</h2>
+      MUSHROOMs
+    </>
+  );
+};
+
+export default ArtistShop;
