+++ client/src/components/Artists/index.js
@@ -3,12 +3,12 @@ import { Outlet } from 'react-router-dom';
 
 const Artists = () => {
   return (
-    <>
+    <div>
       <h2>this is the artist's page</h2>
       <h4>outlet:</h4>
 
       <Outlet />
-    </>
+    </div>
   );
 };
