+++ client/src/pages/Shop.js
@@ -3,13 +3,13 @@ import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useQuery, useMutation } from '@apollo/client';
 import FileBase64 from 'react-file-base64';
-import { QUERY_USER_MERCH } from '../../utils/queries';
-import { ADD_MERCH, REMOVE_MERCH } from '../../utils/mutations';
-import Auth from '../../utils/auth';
+import { QUERY_USER_MERCH } from '../utils/queries';
+import { ADD_MERCH, REMOVE_MERCH } from '../utils/mutations';
+import Auth from '../utils/auth';
 
-import ShopItems from '../ShopItems';
+import Merch from '../components/Merch';
 
-const Store = () => {
+const Shop = () => {
   // console.log('load shop');
   const { loading, data } = useQuery(QUERY_USER_MERCH, {
     variables: { username: Auth.getProfile().data.username }
@@ -178,7 +178,7 @@ const Store = () => {
           ) : (
             merch.map((item) => (
               <div key={item._id} className="merch-container">
-                <ShopItems shopItems={item} />
+                <Merch merch={item} />
                 <button
                   type="button"
                   merchid={item._id}
@@ -195,4 +195,4 @@ const Store = () => {
   );
 };
 
-export default Store;
+export default Shop;
